feat: add `svid-extractor`
Move `extract_leaf_certificate` and `extract_spiffe_id` from
integration tests to lib under `svid-extractor` feature.
